Revamp front page
This removes most of the 'marketing' type content that I dont think
really worked.
I'd like to add more content (eg maybe for the HUD) but I think it might
be better to cut it back for now and then gradually add content that we
are really happy with.
Thoughts?

Note that I have deliberately not removed data associated with the
deleted home page elements to make it easier to add them back in, but
can remove them if people prefer that approach.

Signed-off-by: Simon Bennetts <psiinon@gmail.com>