Allow to link to source content
Allow pages to opt in to link to source (for edition), a link is shown
in the footer. The pages can override the label, base URL, and file path
(e.g. for help pages which live in other repos).
Enable for FAQ, guides, and Getting Started.

Part of #24 - Edit me on GitHub

Signed-off-by: thc202 <thc202@gmail.com>