Add eclipse docs
The plan is to have one page per popular IDE - we dont have to write all
of them, we can ask the community for their contributions.

This page should be a template for the others, so dont hold back on the
feedback ;)

The one thing that I think would be really useful is to explain how to
debug add-ons in each IDE. Anything else you can think of that we should
add asap?

Signed-off-by: Simon Bennetts <psiinon@gmail.com>