Auth docs - prev/next links
Add previous / next links to the auth docs as they are designed to be
read in order.
The links are maintained manually - I'm sure hugo has a cleaver way of
doing this automatically but I couldnt work out how, and these pages
wont change often.
Also removed the ZAP in Ten link which no longer works.

Signed-off-by: Simon Bennetts <psiinon@gmail.com>