Add Roles and Governance documentation
To be used for OpenSSF Best Practices Silver.
https://bestpractices.coreinfrastructure.org/en/projects/24?criteria_level=1

"The project MUST clearly define and document its project governance
model (the way it makes decisions, including key roles)."

"The project MUST clearly define and publicly document the key roles in
the project and their responsibilities, including any tasks those roles
must perform. It MUST be clear who has which role(s), though this might
not be documented in the same way."

Signed-off-by: kingthorin <kingthorin@users.noreply.github.com>